# openapi.yml
openapi: 3.0.3
info:
  title: Ledger V3 POC API
  description: API for the Ledger V3 Proof of Concept with Raft consensus
  version: 1.0.0
  contact:
    name: Formance
    url: https://formance.com
servers:
  - url: http://localhost:9000
    description: Local development server
...
paths:
...
  /{ledgerName}:
    get:
      summary: Get a ledger
      description: Retrieves a ledger by its name.
      operationId: getLedger
      tags:
        - Ledgers
      parameters:
        - name: ledgerName
          in: path
          required: true
          schema:
            type: string
          description: Name of the ledger to retrieve
      responses:
        '200':
          description: Ledger retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLedgerResponse'
        '404':
          description: Ledger not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new ledger
      description: |
        Creates a new ledger with the specified name, drivers, and configurations. Both logStoreDriver and runtimeStoreDriver are required.
        Available drivers: sqlite-mattn (github.com/mattn/go-sqlite3), sqlite-modern (modernc.org/sqlite), and pebble (github.com/cockroachdb/pebble).
        Each store (log store and runtime store) can have its own driver and configuration, allowing for flexible storage setups.
        Each ledger has its own Raft group for data consistency.
      operationId: createLedger
      tags:
        - Ledgers
      parameters:
        - name: ledgerName
          in: path
          required: true
          schema:
            type: string
          description: Name of the ledger to create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLedgerRequest'
      responses:
        '201':
          description: Ledger created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateLedgerResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Ledger already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a ledger
      description: Deletes a ledger by its name. This operation removes the ledger and its associated Raft group from the cluster.
      operationId: deleteLedger
      tags:
        - Ledgers
      parameters:
        - name: ledgerName
          in: path
          required: true
          schema:
            type: string
          description: Name of the ledger to delete
      responses:
        '204':
          description: Ledger deleted successfully
        '404':
          description: Ledger not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /{ledgerName}/raft/state:
    get:
      summary: Get ledger Raft cluster state
      description: Returns the current state of the Raft cluster for the specified ledger, including the list of nodes, the current leader, and the ledger state
      operationId: getLedgerRaftState
      tags:
        - Ledgers
      parameters:
        - name: ledgerName
          in: path
          required: true
          schema:
            type: string
          description: Name of the ledger
      responses:
        '200':
          description: Ledger Raft cluster state retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerClusterStateResponse'
        '404':
          description: Ledger not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /{ledgerName}/transactions:
    post:
      summary: Create a new transaction
      description: Creates a new transaction in the specified ledger with the specified postings
      operationId: createTransaction
      tags:
        - Transactions
      parameters:
        - name: ledgerName
          in: path
          required: true
          schema:
            type: string
          description: Name of the ledger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTransactionResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /{ledgerName}/accounts/{address}/metadata:
    post:
      summary: Save account metadata
      description: Saves metadata for a specific account in the ledger
      operationId: saveAccountMetadata
      tags:
        - Accounts
      parameters:
        - name: ledgerName
          in: path
          required: true
          schema:
            type: string
          description: Name of the ledger
        - name: address
          in: path
          required: true
          schema:
            type: string
          description: Account address
        - name: dryRun
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: If true, the operation is validated but not executed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
              description: Metadata key-value pairs
              example:
                account_type: asset
                label: Bank Account
      responses:
        '204':
          description: Account metadata saved successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /{ledgerName}/_bulk:
    post:
      summary: Bulk operations
      description: Execute multiple operations (create transactions, add metadata, revert transactions, delete metadata) in a single request
      operationId: bulkOperations
      tags:
        - Transactions
      parameters:
        - name: ledgerName
          in: path
          required: true
          schema:
            type: string
          description: Name of the ledger
        - name: continueOnFailure
          in: query
          schema:
            type: boolean
            default: false
          description: Continue processing remaining operations even if one fails
        - name: atomic
          in: query
          schema:
            type: boolean
            default: false
          description: Execute all operations atomically (not yet supported)
        - name: parallel
          in: query
          schema:
            type: boolean
            default: false
          description: Execute operations in parallel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/BulkElement'
      responses:
        '200':
          description: Bulk operations completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request entity too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cluster/state:
    get:
      summary: Get cluster state
      description: Returns the current state of the Raft cluster, including the list of nodes and the current leader
      operationId: getClusterState
      tags:
        - Cluster
      responses:
        '200':
          description: Cluster state retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterStateResponse'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  responses:
    ServiceUnavailable:
      description: Service unavailable (no leader available)
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying the request
          schema:
            type: integer
            format: int32
            minimum: 1
            example: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CreateLedgerRequest:
      type: object
      required:
        - logStoreDriver
        - runtimeStoreDriver
      properties:
        logStoreDriver:
          type: string
          enum:
          - sqlite-modern
          - sqlite-mattn
          - pebble
          description: |
            Log store driver name. 
            Available drivers: sqlite-mattn (github.com/mattn/go-sqlite3), sqlite-modern (modernc.org/sqlite), 
            and pebble (github.com/cockroachdb/pebble)
        runtimeStoreDriver:
          type: string
          enum:
          - sqlite-modern
          - sqlite-mattn
          - pebble
          description: |
            Runtime store driver name. 
            Available drivers: sqlite-mattn (github.com/mattn/go-sqlite3), sqlite-modern (modernc.org/sqlite), 
            and pebble (github.com/cockroachdb/pebble)
        logStoreConfig:
          oneOf:
            - $ref: '#/components/schemas/SQLiteMattnConfig'
            - $ref: '#/components/schemas/SQLiteModernConfig'
            - $ref: '#/components/schemas/PebbleConfig'
          description: Log store driver-specific configuration (optional - DSN/data directory is auto-generated for all drivers)
        runtimeStoreConfig:
          oneOf:
            - $ref: '#/components/schemas/SQLiteMattnConfig'
            - $ref: '#/components/schemas/SQLiteModernConfig'
            - $ref: '#/components/schemas/PebbleConfig'
          description: Runtime store driver-specific configuration (optional - DSN/data directory is auto-generated for all drivers)
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Optional metadata for the ledger
        snapshotThreshold:
          type: integer
          format: int64
          description: Number of logs before triggering a snapshot (optional, uses global config if not set)
    CreateLedgerResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/LedgerInfo'
    GetLedgerResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/LedgerInfo'
...
    LedgerInfo:
      type: object
      required:
        - id
        - name
        - logStoreDriver
        - runtimeStoreDriver
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: Sequential ID for the ledger
        name:
          type: string
          description: Name of the ledger
        logStoreDriver:
          type: string
          enum: [sqlite-mattn, sqlite-modern, pebble]
          description: |
            Log store driver name. 
            Available drivers: sqlite-mattn (github.com/mattn/go-sqlite3), sqlite-modern (modernc.org/sqlite), 
            and pebble (github.com/cockroachdb/pebble)
        runtimeStoreDriver:
          type: string
          enum: [sqlite-mattn, sqlite-modern, pebble]
          description: |
            Runtime store driver name. 
            Available drivers: sqlite-mattn (github.com/mattn/go-sqlite3), sqlite-modern (modernc.org/sqlite), 
            and pebble (github.com/cockroachdb/pebble)
        logStoreConfig:
          oneOf:
            - $ref: '#/components/schemas/SQLiteMattnConfig'
            - $ref: '#/components/schemas/SQLiteModernConfig'
            - $ref: '#/components/schemas/PebbleConfig'
          description: Log store driver-specific configuration
        runtimeStoreConfig:
          oneOf:
            - $ref: '#/components/schemas/SQLiteMattnConfig'
            - $ref: '#/components/schemas/SQLiteModernConfig'
            - $ref: '#/components/schemas/PebbleConfig'
          description: Runtime store driver-specific configuration
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Metadata for the ledger
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601 format)
        snapshotThreshold:
          type: integer
          format: int64
          description: Number of logs before triggering a snapshot (0 means use global config)
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Deletion timestamp (always null with hard delete)
        status:
          type: string
          enum: [Active, Deleted]
          description: Ledger status
    CreateTransactionRequest:
      type: object
      properties:
        accountMetadata:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: string
          description: Metadata for accounts involved in the transaction
        timestamp:
          type: string
          format: date-time
          description: Transaction timestamp (ISO 8601 format). If not provided, current time is used.
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Transaction metadata
        reference:
          type: string
          description: Transaction reference
        postings:
          type: array
          items:
            $ref: '#/components/schemas/PostingRequest'
          description: List of postings for the transaction (required if script is not provided)
        script:
          $ref: '#/components/schemas/Script'
          description: Numscript script to execute (required if postings is not provided). Cannot be used together with postings.
        dryRun:
          type: boolean
          default: false
          description: If true, validate the transaction without applying it
        idempotencyKey:
          type: string
          description: Idempotency key to prevent duplicate transactions
...
    Script:
      type: object
      required:
        - plain
      properties:
        plain:
          type: string
          description: The numscript source code
        vars:
          type: object
          additionalProperties:
            type: string
          description: Variables to pass to the script
    BulkElement:
      type: object
      required:
        - action
        - data
      properties:
        action:
          type: string
          enum:
            - CREATE_TRANSACTION
            - ADD_METADATA
            - REVERT_TRANSACTION
            - DELETE_METADATA
          description: The type of operation to perform
        ik:
          type: string
          description: Idempotency key for this operation
        data:
          oneOf:
            - $ref: '#/components/schemas/CreateTransactionRequest'
            - $ref: '#/components/schemas/AddMetadataRequest'
            - $ref: '#/components/schemas/RevertTransactionRequest'
            - $ref: '#/components/schemas/DeleteMetadataRequest'
          description: Operation-specific data
...
    BulkResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BulkResult'
        errorCode:
          type: string
        errorMessage:
          type: string
...
    CreateTransactionResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - transaction
          properties:
            transaction:
              $ref: '#/components/schemas/TransactionResponse'
            accountMetadata:
              type: object
              additionalProperties:
                type: object
                additionalProperties:
                  type: string
              description: Metadata for accounts involved in the transaction
...
    ClusterStateResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - state
            - localNode
            - innerState
          properties:
            state:
              type: string
              enum: [Leader, Follower, Candidate, PreCandidate, Unknown]
              description: Current state of the local node
            leader:
              type: integer
              format: int64
              minimum: 0
              description: ID of the current leader (0 if no leader)
            localNode:
              type: integer
              format: int64
              minimum: 0
              description: ID of the local node
            nodes:
              type: array
              items:
                $ref: '#/components/schemas/NodeInfo'
              description: List of all nodes in the cluster
            raftStatus:
              $ref: '#/components/schemas/RaftStatus'
              description: Complete Raft status information
            innerState:
              $ref: '#/components/schemas/SystemState'
              description: State contained in the FSM
    LedgerClusterStateResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - state
            - localNode
            - innerState
          properties:
            state:
              type: string
              enum: [Leader, Follower, Candidate, PreCandidate, Unknown]
              description: Current state of the local node
            leader:
              type: integer
              format: int64
              minimum: 0
              description: ID of the current leader (0 if no leader)
            localNode:
              type: integer
              format: int64
              minimum: 0
              description: ID of the local node
            nodes:
              type: array
              items:
                $ref: '#/components/schemas/NodeInfo'
              description: List of all nodes in the cluster
            raftStatus:
              $ref: '#/components/schemas/RaftStatus'
              description: Complete Raft status information
            innerState:
              $ref: '#/components/schemas/LedgerState'
              description: State contained in the FSM
...
    NodeInfo:
      type: object
      required:
        - id
        - address
        - suffrage
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
          description: Node ID
        address:
          type: string
          description: Node address
        suffrage:
          type: string
          enum: [Voter, Nonvoter, Learner]
          description: Node suffrage (Voter, Nonvoter, or Learner)
...
    RaftStatus:
      type: object
      required:
        - state
        - term
        - leader
        - applied
        - commit
        - lastIndex
        - vote
      properties:
        state:
          type: string
          enum: [Leader, Follower, Candidate, PreCandidate, Unknown]
          description: Current Raft state
        term:
          type: integer
          format: int64
          minimum: 0
          description: Current term
        leader:
          type: integer
          format: int64
          minimum: 0
          description: ID of the current leader (0 if no leader)
        applied:
          type: integer
          format: int64
          minimum: 0
          description: Index of the last applied entry
        commit:
          type: integer
          format: int64
          minimum: 0
          description: Index of the last committed entry
        lastIndex:
          type: integer
          format: int64
          minimum: 0
          description: Index of the last entry in the log
        vote:
          type: integer
          format: int64
          minimum: 0
          description: ID of the node that received the vote (0 if no vote)
        progress:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ProgressInfo'
          description: Progress information for each node in the cluster
...
    SQLiteMattnConfig:
      type: object
      description: SQLite Mattn driver configuration (github.com/mattn/go-sqlite3) - empty object, DSN is automatically generated
      properties: {}
      example: {}
    SQLiteModernConfig:
      type: object
      description: SQLite Modern driver configuration (modernc.org/sqlite) - empty object, DSN is automatically generated
      properties: {}
      example: {}
    PebbleConfig:
      type: object
      description: Pebble driver configuration (github.com/cockroachdb/pebble) - empty object, data directory is automatically generated
      properties: {}
      example: {}
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          description: Error code identifier
        errorMessage:
          type: string
          description: Human-readable error message
...